Tempomat_2 library updates; Tempomat_WithFeedback created - V1.0 established
Tempomat_WithFeedback - a tachometer with a feedback info of the acceleration pedal position due to added potentiometer mechanically in the vehicle.
The system calculates the error of the revolutions per minute and steers the acceleration by moving the acceleration pedal. The dedicated task constantly (10ms period) checks if the gas position is equal to the one desired and calculated in the main Task, where a PID controller calculates the steering values.